Make puma an optional dependency
Metatron is a Rack comforming application, which means it will actually
work with any Rack compatible server. Making it a runtime dependency is
unecessary and will likely just bite us later when a downstream
application requires us to bump the puma requirement.

Instead, let's remove it as a runtime dependency. To ensure that the
example in the README can still run with puma, it was added to the
Gemfile as an example of a "rack compatible server".