Rackify Controller and CompositeController
This removes Sinatra as the underlying class of Controller and
CompositeController and replaces features previously implemented by
Sinatra with plain Rack.

The most notable feature kept is ETags/Conditional Get which is kept
using the If-None-Match header.

An additional thing to note is that Controller is implemented slightly
differently than PingController. PingController does not really carry
any state so it's implemented as a persistent instance that can be
called. Metatron::Controller does want to carry some state that can be
used in the implementing class's methods, so intead of a single instance
the Controller will be instantiated per request when called.