Remove dependency on Sinatra
Now that all of the controllers no longer use Sinatra, we can replace
it's requirement with Rack. Additionally, we can start testing against
Rack 3 since Sinatra was preventing that from happening. In a future
commit, we could investigate setting up CI for both versions to ensure
we don't break either.

This also removes the require for base64 since it will be [promoted][1]
from a standard gem to a bundled gem in Ruby 3.3 (which means we can't
require it without it being a dependency or in a Gemfile)

[1]: https://bugs.ruby-lang.org/issues/19351